# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для доступа к расписанию МПК ТИУ",
        "title": "Raspishika API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/departments": {
            "get": {
                "description": "Returns list of all existing departments",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedule",
                    "departments"
                ],
                "summary": "Get departments",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_azzimoda_raspishika-gx_internal_model.Department"
                            }
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_azzimoda_raspishika-gx_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/groups": {
            "get": {
                "description": "Returns list of all existing groups, optionally filtered by department",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedule",
                    "groups"
                ],
                "summary": "Get groups",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Department name",
                        "name": "department",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_azzimoda_raspishika-gx_internal_model.Group"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_azzimoda_raspishika-gx_internal_model.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_azzimoda_raspishika-gx_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/groups/{name}": {
            "get": {
                "description": "Returns a group by name",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Get group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_azzimoda_raspishika-gx_internal_model.Group"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_azzimoda_raspishika-gx_internal_model.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_azzimoda_raspishika-gx_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/schedule": {
            "get": {
                "description": "Returns schedule of a group or a teacher",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedule",
                    "groups",
                    "teachers"
                ],
                "summary": "Get schedule",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group name",
                        "name": "group",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Teacher name or college internal ID",
                        "name": "teacher",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_azzimoda_raspishika-gx_internal_model.ScheduleData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_azzimoda_raspishika-gx_internal_model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_azzimoda_raspishika-gx_internal_model.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_azzimoda_raspishika-gx_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/teachers": {
            "get": {
                "description": "Returns list of all existing teachers",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedule",
                    "teachers"
                ],
                "summary": "Get teachers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_azzimoda_raspishika-gx_internal_model.Teacher"
                            }
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_azzimoda_raspishika-gx_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/teachers/search": {
            "get": {
                "description": "Returns a list of teachers matched by part of name",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teachers"
                ],
                "summary": "Search teachers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Teacher name part",
                        "name": "q",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_azzimoda_raspishika-gx_internal_model.Teacher"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_azzimoda_raspishika-gx_internal_model.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_azzimoda_raspishika-gx_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/teachers/{name_or_id}": {
            "get": {
                "description": "Returns a teacher by name or college internal ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teachers"
                ],
                "summary": "Get teacher",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Teacher name or college internal ID",
                        "name": "name_or_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_azzimoda_raspishika-gx_internal_model.Teacher"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_azzimoda_raspishika-gx_internal_model.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_azzimoda_raspishika-gx_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_azzimoda_raspishika-gx_internal_model.Department": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "28728"
                },
                "name": {
                    "type": "string",
                    "example": "АиЭС"
                },
                "year": {
                    "type": "integer",
                    "example": 2026
                }
            }
        },
        "github_com_azzimoda_raspishika-gx_internal_model.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "group not found"
                }
            }
        },
        "github_com_azzimoda_raspishika-gx_internal_model.Group": {
            "type": "object",
            "properties": {
                "department_id": {
                    "type": "string",
                    "example": "15"
                },
                "department_name": {
                    "type": "string",
                    "example": "Отделение СОНХ"
                },
                "group_id": {
                    "type": "string",
                    "example": "205"
                },
                "group_name": {
                    "type": "string",
                    "example": "ИСПт-22-(9)-2"
                },
                "year": {
                    "type": "integer",
                    "example": 2026
                }
            }
        },
        "github_com_azzimoda_raspishika-gx_internal_model.Pair": {
            "type": "object",
            "properties": {
                "classroom": {
                    "type": "string",
                    "example": "215"
                },
                "discipline": {
                    "type": "string",
                    "example": "Математика"
                },
                "end_time": {
                    "type": "string",
                    "example": "9:45"
                },
                "group": {
                    "type": "string",
                    "example": "ИСПт-22-(9)-2"
                },
                "kind": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_azzimoda_raspishika-gx_internal_model.PairKind"
                        }
                    ],
                    "example": "subject"
                },
                "label": {
                    "type": "string",
                    "example": "Математика"
                },
                "number": {
                    "type": "integer",
                    "example": 1
                },
                "replaced": {
                    "type": "boolean"
                },
                "start_time": {
                    "type": "string",
                    "example": "8:00"
                },
                "subgroup": {
                    "type": "string",
                    "example": "1"
                },
                "teacher": {
                    "type": "string",
                    "example": "Иванов Иван Иванович"
                },
                "title": {
                    "type": "string",
                    "example": "Консультация"
                }
            }
        },
        "github_com_azzimoda_raspishika-gx_internal_model.PairKind": {
            "type": "string",
            "enum": [
                "empty",
                "vacation",
                "event",
                "session",
                "iga",
                "subject",
                "practice",
                "exam",
                "consultation"
            ],
            "x-enum-varnames": [
                "PairKindEmpty",
                "PairKindVacation",
                "PairKindEvent",
                "PairKindSession",
                "PairKindIGA",
                "PairKindSubject",
                "PairKindPractice",
                "PairKindExam",
                "PairKindConsultation"
            ]
        },
        "github_com_azzimoda_raspishika-gx_internal_model.ScheduleConfig": {
            "type": "object",
            "properties": {
                "group": {
                    "$ref": "#/definitions/github_com_azzimoda_raspishika-gx_internal_model.Group"
                },
                "is_dark": {
                    "type": "boolean"
                },
                "teacher": {
                    "$ref": "#/definitions/github_com_azzimoda_raspishika-gx_internal_model.Teacher"
                }
            }
        },
        "github_com_azzimoda_raspishika-gx_internal_model.ScheduleData": {
            "type": "object",
            "properties": {
                "config": {
                    "$ref": "#/definitions/github_com_azzimoda_raspishika-gx_internal_model.ScheduleConfig"
                },
                "days": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_azzimoda_raspishika-gx_internal_model.ScheduleDay"
                    }
                },
                "is_old": {
                    "type": "boolean"
                }
            }
        },
        "github_com_azzimoda_raspishika-gx_internal_model.ScheduleDay": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string",
                    "example": "2026-09-01"
                },
                "pairs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_azzimoda_raspishika-gx_internal_model.Pair"
                    }
                },
                "week_day": {
                    "type": "string",
                    "example": "вторник"
                },
                "week_kind": {
                    "type": "string",
                    "example": "нечетная"
                }
            }
        },
        "github_com_azzimoda_raspishika-gx_internal_model.Teacher": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Иванов Иван Иванович"
                },
                "teacher_id": {
                    "type": "string",
                    "example": "205"
                }
            }
        }
    }
}